Add fetch Helper file and moved fetch calls from App to helper file.
